requirements.txt
fastapi
uvicorn
pydantic
httpx
motor
redis
python-dotenv
